Tweaks for a clean compile with gcc -ansi -Wall -pedantic

+++ hdf/util/hdfrseq.c
@@ -528,7 +528,7 @@ int rimage(int usepal)
 *  Open the window with the W command
 */
 
-(void)printf("\033^W;%d;%d;%d;%d;0;rseq^",xwhere,ywhere,xdim*xfact,ydim*yfact);
+(void)printf("\033^W;%d;%d;%ld;%ld;0;rseq^",xwhere,ywhere,(long)(xdim*xfact),(long)(ydim*yfact));
 
 /*
 *  If a palette should be used, send it with the M command.
